Refactor: Centralize WebSocket test setup in WebSocketIntegrationTestsBase
